+++ packages/picker-button/src/mobile/index.ts
@@ -1,1 +1,2 @@
-export { PickerButtonMobile, PickerButtonMobileProps } from './Component.mobile';
+export { PickerButtonMobile } from './Component.mobile';
+export type { PickerButtonMobileProps } from './Component.mobile';