+++ packages/stack/src/component.test.tsx
@@ -1,8 +1,8 @@
 import React, { Fragment } from 'react';
 import { render } from '@testing-library/react';
 
-import { Stack } from '.';
-import { stackingOrder } from '@alfalab/stack-context';
+import { Stack } from '@alfalab/core-components-stack';
+import { stackingOrder } from '@alfalab/core-components-stack-context';
 
 const getZIndex = (element: HTMLElement) =>
     element ? parseInt(element.style.getPropertyValue('z-index'), 10) : 0;