+++ packages/spinner/CHANGELOG.md
@@ -1,4 +1,4 @@
-# Change Log
+# @alfalab/core-components-spinner
 
 ## 4.0.9